+++ src/soapbox/actions/__tests__/announcements.test.ts
@@ -4,7 +4,7 @@ import { List as ImmutableList } from 'immutable';
 import { fetchAnnouncements, dismissAnnouncement, addReaction, removeReaction } from 'soapbox/actions/announcements';
 import { __stub } from 'soapbox/api';
 import { mockStore, rootState } from 'soapbox/jest/test-helpers';
-import { normalizeAnnouncement, normalizeInstance } from 'soapbox/normalizers';
+import { normalizeAnnouncement } from 'soapbox/normalizers';
 
 import type { APIEntity } from 'soapbox/types/entities';