+++ src/soapbox/actions/__tests__/suggestions.test.ts
@@ -3,7 +3,6 @@ import { Map as ImmutableMap } from 'immutable';
 
 import { __stub } from 'soapbox/api';
 import { mockStore, rootState } from 'soapbox/jest/test-helpers';
-import { normalizeInstance } from 'soapbox/normalizers';
 
 import {
   fetchSuggestions,