+++ src/soapbox/features/ads/components/ad.tsx
@@ -76,7 +76,7 @@ const Ad: React.FC<IAd> = ({ ad }) => {
       <Card className='py-4' variant='rounded'>
         <Stack space={4}>
           <HStack alignItems='center' space={3}>
-            <Avatar src={instance.thumbnail} size={42} />
+            <Avatar src={instance.thumbnail.url} size={42} />
 
             <Stack grow>
               <HStack space={1}>