+++ src/soapbox/features/auth-login/components/__tests__/login-form.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { normalizeInstance } from 'soapbox/normalizers';
+import { instanceSchema } from 'soapbox/schemas';
 
 import { fireEvent, render, screen } from '../../../../jest/test-helpers';
 import LoginForm from '../login-form';
@@ -9,7 +9,7 @@ describe('<LoginForm />', () => {
   it('renders for Pleroma', () => {
     const mockFn = jest.fn();
     const store = {
-      instance: normalizeInstance({
+      instance: instanceSchema.parse({
         version: '2.7.2 (compatible; Pleroma 2.3.0)',
       }),
     };
@@ -22,7 +22,7 @@ describe('<LoginForm />', () => {
   it('renders for Mastodon', () => {
     const mockFn = jest.fn();
     const store = {
-      instance: normalizeInstance({
+      instance: instanceSchema.parse({
         version: '3.0.0',
       }),
     };