+++ src/soapbox/jest/mock-stores.tsx
@@ -1,26 +1,26 @@
 import { fromJS } from 'immutable';
 
 import alexJson from 'soapbox/__fixtures__/pleroma-account.json';
-import { normalizeInstance } from 'soapbox/normalizers';
+import { instanceSchema } from 'soapbox/schemas';
 
 import { buildAccount } from './factory';
 
 /** Store with registrations open. */
-const storeOpen = { instance: normalizeInstance({ registrations: true }) };
+const storeOpen = { instance: instanceSchema.parse({ registrations: true }) };
 
 /** Store with registrations closed. */
-const storeClosed = { instance: normalizeInstance({ registrations: false }) };
+const storeClosed = { instance: instanceSchema.parse({ registrations: false }) };
 
 /** Store with registrations closed, and Pepe enabled & open. */
 const storePepeOpen = {
-  instance: normalizeInstance({ registrations: false }),
+  instance: instanceSchema.parse({ registrations: false }),
   soapbox: fromJS({ extensions: { pepe: { enabled: true } } }),
   verification: { instance: fromJS({ registrations: true }) },
 };
 
 /** Store with registrations closed, and Pepe enabled & closed. */
 const storePepeClosed = {
-  instance: normalizeInstance({ registrations: false }),
+  instance: instanceSchema.parse({ registrations: false }),
   soapbox: fromJS({ extensions: { pepe: { enabled: true } } }),
   verification: { instance: fromJS({ registrations: false }) },
 };