+++ src/soapbox/reducers/instance.ts
@@ -3,7 +3,6 @@ import { Map as ImmutableMap, List as ImmutableList } from 'immutable';
 
 import { ADMIN_CONFIG_UPDATE_REQUEST, ADMIN_CONFIG_UPDATE_SUCCESS } from 'soapbox/actions/admin';
 import { PLEROMA_PRELOAD_IMPORT } from 'soapbox/actions/preload';
-import { normalizeInstance } from 'soapbox/normalizers';
 import { type Instance, instanceSchema } from 'soapbox/schemas';
 import KVStore from 'soapbox/storage/kv-store';
 import { ConfigDB } from 'soapbox/utils/config-db';
@@ -18,7 +17,7 @@ import type { AnyAction } from 'redux';
 const initialState: Instance = instanceSchema.parse({});
 
 const importInstance = (_state: typeof initialState, instance: any) => {
-  return normalizeInstance(instance);
+  return instanceSchema.parse(instance);
 };
 
 const preloadImport = (state: typeof initialState, action: Record<string, any>, path: string) => {
@@ -46,8 +45,8 @@ const importConfigs = (state: typeof initialState, configs: ImmutableList<any>) 
       const registrationsOpen = getConfigValue(value, ':registrations_open') as boolean | undefined;
       const approvalRequired = getConfigValue(value, ':account_approval_required') as boolean | undefined;
 
-      draft.registrations = registrationsOpen ?? draft.registrations;
-      draft.approval_required = approvalRequired ?? draft.approval_required;
+      if (typeof registrationsOpen !== 'undefined') draft.registrations.enabled = registrationsOpen;
+      if (typeof approvalRequired !== 'undefined') draft.registrations.approval_required = approvalRequired;
     }
 
     if (simplePolicy) {